DdjangoStocksite/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
comtypes==1.1.10
constants==0.6.0
creon==0.1
cryptography==3.4.7
dataclasses==0.8
Django==3.2.5
django-baton==2.2.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djongo==1.3.6
MouseInfo==0.1.3
node==0.9.26
nodeenv==1.6.0
odict==1.7.0
Pillow==8.3.1
plumber==1.6
psutil==5.7.3
PyAutoGUI==0.9.53
pycparser==2.20
PyGetWindow==0.0.9
PyJWT==1.7.1
pymongo==3.12.0
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
pytz==2021.1
pywin32==228
pywinauto==0.6.8
six==1.16.0
sqlparse==0.2.4
typing-extensions==3.7.4.3
wincertstore==0.2
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.0.1
zope.interface==5.4.0
zope.lifecycleevent==4.3
